+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Notes
 
+- Use Kotlin Duration and TimeSource
 - Switch to single-project build
 - Add `currentTime` variable to decrease code repetitions
 - Rename `CallbackCaller` to `CallbackHandler`